requirements.txt
fastapi
uvicorn
langchain
langchain-community
ollama
pytesseract
python-multipart
whisper
PyMuPDF
gTTS
Pillow
